feat: make `timeout` for inferencing calls configurable (#65)
* feat: make `timeout` for inferencing calls configurable

* chore: documentation

* chore: useful comments

* chore: format

* feat: add `TIMEOUT` to all notebooks

* chore: debug

* chore: debug

* chore: debugging

---------

+++ tests/term_analysis/test_keyword_analyzer.py
@@ -70,7 +70,7 @@ class TestKeywordAnalyzer(unittest.TestCase):
             lang_util=self.lang_util,
         )
         unique_counts = len(counts.index.get_level_values(0).unique())
-        actual_labels_shown = np.int(np.ceil(30 / unique_counts)) * unique_counts
+        actual_labels_shown = np.int_(np.ceil(30 / unique_counts)) * unique_counts
         self.assertEqual(
             len(top_counts) == actual_labels_shown, True, "Key word analyzer test fails"
         )
